Build libjsonnet.js using emscripten and remove static copy.
