+++ test/test_err.js
@@ -1,4 +1,4 @@
-var Jsonnet = require("../index");
+var Jsonnet = require("../lib/index");
 var jsonnet = new Jsonnet();
 var fs = require('fs');
 var path = require('path');
@@ -8,7 +8,7 @@ describe('jsonnet', function(){
   describe('#eval() error', function(){
     it('should catch error and include foo string', function(){
       var filepath = path.join(__dirname, 'src' , 'error.jsonnet');
-      var code = fs.readFileSync(filepath);
+      var code = fs.readFileSync(filepath).toString();
       try {
         jsonnet.eval(code);
       } catch(e) {
